Ctrax:
* When reading compressed AVIs with a keyframe in every frame, the 
buffer offset was not being set correctly initially (initialized to 1 
instead of 0, which caused trouble when the length of the buffer was 1). 
Fixed this bug. 

behavioralmicroarray:

* Fixed bug in make_ctrax_result_movie. If the movie was an AVI, then 
fid was being set to -1, and make_ctrax_result_movie was interpreting 
this as the movie being unreadable. Also, it tried to close the fid at 
the end of the function. 

matlab:

* Changed get_readframe_fcn so that it returns fid=0 when using mmreader 
to read an AVI.